% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_metric}
\alias{append_metric}
\title{Append metric failure entries to an existing metrics list}
\usage{
append_metric(m, samples, exons, type, details, genes)
}
\arguments{
\item{m}{Existing metrics list.}

\item{samples}{Vector of sample names.}

\item{exons}{Vector of exon identifiers.}

\item{type}{Metric type label.}

\item{details}{Vector of detail messages.}

\item{genes}{Vector of gene names.}
}
\value{
A list containing updated metric fields.
}
\description{
Append metric failure entries to an existing metrics list
}
