% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo-package.R
\docType{package}
\name{echo-package}
\alias{echo-package}
\alias{echo_package}
\title{echo: Echo Code Evaluations}
\description{
Capture code evaluations and script executions by expressions, outputs, and condition calls for logging.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmbarbone/echo}
  \item \url{https://jmbarbone.github.io/echo/}
  \item Report bugs at \url{https://github.com/jmbarbone/echo/issues}
}

}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
