% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R, R/print.echo.R, R/summary.echo.R
\name{echo}
\alias{echo}
\alias{print.echo}
\alias{summary.echo}
\title{ECHO}
\usage{
echo(
  tree,
  n_samples,
  latent_duration,
  infectious_duration,
  Re,
  time_scale = "years",
  CI = 0.95
)

\method{print}{echo}(x, ...)

\method{summary}{echo}(x, credible_interval = 0.95, ...)
}
\arguments{
\item{tree}{The phylogenetic tree(s) for which ECHO will estimate cryptic case counts.
Can be an object of class \code{phylo}, \code{treedata}, a string indicating the filepath of a Newick file, or a numeric value equal to the total branch length of the tree.
If you want \code{echo} to process a posterior sample of trees, input the sample either as a list or a vector.}

\item{n_samples}{The number of samples. This should be equal to the number of tips in the phylogeny.}

\item{latent_duration}{The average duration of the latent period in days.}

\item{infectious_duration}{The average duration of the infectious period in days.}

\item{Re}{The effective reproduction number. If unknown, provide an \code{NA} value; in this case ECHO-C will not be computed.}

\item{time_scale}{String indicating the time scale that \code{tree} is measured in.
Can be either \code{"days"} or \code{"years"}. Default value is \code{"years"}.}

\item{CI}{The desired coverage probability of the confidence intervals. Must be a single value between 0 and 1. Default value is 0.95.}

\item{x}{Object of class \code{echo}.}

\item{credible_interval}{Level of the credible intervals. Only applicable if \code{n_trees>1}.}
}
\value{
An object of class \code{echo}. Contains the following elements:
\itemize{
\item \code{n_trees}: The number of input trees.
\item \code{n_samples}: The number of samples, equal to the number of tips.
\item \code{A}: A list containing the point estimate(s) and confidence interval(s) corresponding to ECHO-A (see Details).
\item \code{B}: A list containing the point estimate(s) and confidence interval(s) corresponding to ECHO-B (see Details).
\item \code{C}: A list containing the point estimate(s) and confidence interval(s) corresponding to ECHO-C (see Details).
}
}
\description{
\code{echo} will estimate the number of cryptic cases related to a phylogenetic tree.
}
\examples{
tree <- ape::read.tree("../data-raw/example_tree.nwk")
result <- echo(tree, n_samples=31, latent_duration=10, infectious_duration=8, Re=2, time_scale="days")

result

summary(result)

}
