% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df.R
\name{extract_df}
\alias{extract_df}
\title{Extract point estimates from ECHO object}
\usage{
extract_df(x)
}
\arguments{
\item{x}{An object of class \code{echo}.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{estimator}: Indicates which of ECHO-A, -B, or -C the row corresponds to
\item \code{point}: The point estimate
\item \code{LB}: The lower bound on the confidence interval
\item \code{UB}: The upper bound on the confidence interval
}
}
\description{
Takes ECHO output and returns a data.frame with the point estimates in an easily manipulable format.
}
