% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_chromosomes}
\alias{filter_chromosomes}
\title{Filter a data frame by chromosome column}
\usage{
filter_chromosomes(df, include = NULL, exclude = NULL)
}
\arguments{
\item{df}{Input data frame.}

\item{include}{Optional vector of chromosomes to include.}

\item{exclude}{Optional vector of chromosomes to exclude.}
}
\value{
Filtered data frame (empty if input is empty).
}
\description{
Filter a data frame by chromosome column
}
