% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_utils.R
\name{flag_terminal_only_calls}
\alias{flag_terminal_only_calls}
\title{Flag CNV calls that affect only terminal exons}
\usage{
flag_terminal_only_calls(cnv_calls, bed_file)
}
\description{
Flag CNV calls that affect only terminal exons
}
