% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Primary value.}

\item{b}{Fallback value.}
}
\description{
Returns the first argument if not NULL, otherwise the second.
}
