% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_utils.R
\name{handle_off_target_regions}
\alias{handle_off_target_regions}
\title{Reclassify off-target / filler BED intervals before exon numbering}
\usage{
handle_off_target_regions(
  df,
  pattern = "^HorsROI",
  handling = c("na", "remove", "merge"),
  verbose = TRUE
)
}
\description{
Reclassify off-target / filler BED intervals before exon numbering
}
