% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{melt_coverage}
\alias{melt_coverage}
\title{Convert wide coverage matrix to long-format data frame}
\usage{
melt_coverage(mat, exon_range)
}
\arguments{
\item{mat}{Coverage matrix.}

\item{exon_range}{Vector of exon indices.}
}
\value{
Long-format data frame with exon, sample, and coverage columns.
}
\description{
Base R replacement for reshape::melt.
}
