% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_utils.R
\name{plot_coverage_pca}
\alias{plot_coverage_pca}
\title{Plot PCA of Sample Coverage Profiles}
\usage{
plot_coverage_pca(
  counts,
  sample_names,
  output_pdf = NULL,
  color_by = NULL,
  scale = TRUE
)
}
\description{
Plot PCA of Sample Coverage Profiles
}
