% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_bed}
\alias{validate_bed}
\title{Validate BED file schema}
\usage{
validate_bed(bed_path)
}
\arguments{
\item{bed_path}{Character string. Path to the BED file.}
}
\value{
Invisibly returns \code{NULL} if validation succeeds.
}
\description{
Validate BED file schema
}
