% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_utils.R
\name{add_within_gene_indices}
\alias{add_within_gene_indices}
\title{Convert global exon indices to within-gene indices}
\usage{
add_within_gene_indices(cnv_calls, bed_file)
}
\description{
Convert global exon indices to within-gene indices
}
