% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_utils.R
\name{compute_gene_gap_positions}
\alias{compute_gene_gap_positions}
\title{Compute gap-inserted x-axis positions for CNV window plots}
\usage{
compute_gene_gap_positions(bed_file, exon_range, gap = 1)
}
\description{
Compute gap-inserted x-axis positions for CNV window plots
}
