% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_utils.R
\name{detect_problematic_exons}
\alias{detect_problematic_exons}
\title{Detect Problematic Exons}
\usage{
detect_problematic_exons(
  count_matrix,
  chromosomes = NULL,
  mad_quantile = 0.9,
  min_mean = 20,
  gc = NULL,
  gc_min = 0.1,
  gc_max = 0.9
)
}
\description{
Detect Problematic Exons
}
