% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{echo}
\alias{echo}
\title{Echo}
\usage{
echo(
  exprs,
  log = echo_get_log(),
  msg = echo_get_msg(),
  level = echo_get_level()
)
}
\arguments{
\item{msg}{Logical, if \code{FALSE} does not output a message}

\item{level}{Sets the echo level (see details)}

\item{expr}{Expressions to evaluate}

\item{to}{A connection or file name for outputs}
}
\value{
returns the last evaluated value, invisibly
}
\description{
Echo expression or a file
}
\details{
Levels of output can be controlled with \code{level}:

\describe{
\item{\code{4}}{\code{EXP}: logs expressions that were evaluated}
\item{\code{3}}{\code{OUT}: logs outputs from expressions}
\item{\code{2}}{\code{MSG}: logs messages}
\item{\code{1}}{\code{WRN}: logs warnings}
\item{\code{0}}{\code{ERR}: logs errors}
}

When set, all outputs at the \code{level} or below are run. Errors are always
logged as they will interrupt and stop the program.
}
\examples{
try(echo({
  1 + 1
  Sys.sleep(2)
  head(mtcars)
  message(1)
  warning(2)
  stop(3)
}))
}
