% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_utils.R
\name{pad_gene_terminal_exons}
\alias{pad_gene_terminal_exons}
\title{Pad the outer edge of each gene's first and last exon}
\usage{
pad_gene_terminal_exons(
  bed_file,
  padding = 0,
  chr_lengths = NULL,
  verbose = TRUE
)
}
\description{
Pad the outer edge of each gene's first and last exon
}
