% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_process_bed.R
\name{process_bed_file}
\alias{process_bed_file}
\title{Process BED File for CNV Analysis}
\usage{
process_bed_file(
  input_bed,
  output_bed,
  bed_process = "STANDARD",
  bed_zero_based = TRUE,
  refseqgene = NULL,
  transcripts_file = NULL,
  unknown_gene = FALSE,
  gene_list_restrict = NULL,
  chr_list_restrict = NULL,
  exon_sep = " ",
  gene_name_collapse = " ",
  customexon = FALSE,
  auto_exon_number = TRUE,
  region_numbering_mode = "bed_text",
  gene_name_keep = NULL,
  list_genes = NULL,
  genes_file = NULL,
  panel_files = NULL,
  genome_version = NULL,
  txdb = NULL,
  gene_field_index = NULL,
  off_target_pattern = "^HorsROI",
  off_target_handling = c("na", "remove", "merge")
)
}
\description{
Process BED File for CNV Analysis
}
