% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bed.R
\name{process_bed_file}
\alias{process_bed_file}
\title{Process BED File for CNV Analysis (RefSeq Select version)}
\usage{
process_bed_file(
  input_bed,
  output_bed,
  bed_process = "STANDARD",
  refseqgene = NULL,
  transcripts_file = NULL,
  unknown_gene = FALSE,
  gene_list_restrict = NULL,
  chr_list_restrict = NULL,
  exon_sep = NULL,
  kmer = NULL,
  customexon = FALSE,
  list_genes = NULL,
  genes_file = NULL,
  genome_version = NULL,
  txdb = NULL,
  gene_field_index = 1
)
}
\description{
REGEN mode now uses a curated RefSeq Select exon BED/GTF-style file
instead of TxDb inference (which is not clinically stable).
}
