% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_filename}
\alias{sanitize_filename}
\title{Sanitize a string for use as a filename}
\usage{
sanitize_filename(name)
}
\arguments{
\item{name}{Character string to sanitize.}
}
\value{
Sanitized filename string.
}
\description{
Sanitize a string for use as a filename
}
