% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_utils.R
\name{stop_if_missing}
\alias{stop_if_missing}
\title{Stop if value is NULL or empty}
\usage{
stop_if_missing(val, msg)
}
\arguments{
\item{val}{Object to validate.}

\item{msg}{Error message if validation fails.}
}
\description{
Stop if value is NULL or empty
}
