% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_not_file}
\alias{stop_if_not_file}
\title{Stop if file does not exist}
\usage{
stop_if_not_file(path, msg)
}
\arguments{
\item{path}{Path to the file.}

\item{msg}{Error message to display if file is missing.}
}
\value{
Invisibly returns \code{NULL}, or stops with an error.
}
\description{
Stop if file does not exist
}
